# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pharmacy POS API",
    "version": "1.0.0",
    "description": "Generated from shared Zod schemas (SSOT)"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
    "/sales": {
      "get": {
        "summary": "List sales with optional search",
        "operationId": "listSales",
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "wildcardSearch",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "success": {
                    "summary": "Sales list",
                    "value": {
                      "success": true,
                      "message": "OK",
                      "data": [
                        {
                          "_id": "64f0a1b2c3d4e5f678901234",
                          "saleNumber": "20250101001",
                          "totalAmount": 1234.56
                        }
                      ],
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a sale",
        "operationId": "createSale",
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "created": {
                    "summary": "Created",
                    "value": {
                      "success": true,
                      "message": "Created",
                      "data": {
                        "_id": "64f0a1b2c3d4e5f678901234",
                        "saleNumber": "20250101001",
                        "totalAmount": 100
                      },
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/customers": {
      "get": {
        "summary": "List customers",
        "operationId": "listCustomers",
        "tags": [
          "Customers"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "customers": {
                    "summary": "Customer list",
                    "value": {
                      "success": true,
                      "message": "OK",
                      "data": [
                        {
                          "_id": "64f0a1b2c3d4e5f678901234",
                          "code": "C00001",
                          "name": "John Doe",
                          "membershipLevel": "regular"
                        }
                      ],
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a customer",
        "operationId": "createCustomer",
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "created": {
                    "summary": "Created",
                    "value": {
                      "success": true,
                      "message": "Created",
                      "data": {
                        "_id": "64f0a1b2c3d4e5f678901234",
                        "code": "C00001",
                        "name": "John Doe"
                      },
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string"
                },
                "param": {
                  "type": "string"
                },
                "location": {
                  "type": "string"
                }
              },
              "required": [
                "msg"
              ]
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "statusCode": {
            "type": "number"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "SaleCreateRequest": {
        "type": "object",
        "properties": {
          "saleNumber": {
            "type": "string"
          },
          "customer": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "product": {
                  "$ref": "#/definitions/SaleCreateRequest/properties/customer"
                },
                "quantity": {
                  "type": "number",
                  "minimum": 0.001,
                  "maximum": 999999
                },
                "price": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 999999.99
                },
                "unitPrice": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 999999.99
                },
                "discount": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 999999.99
                },
                "subtotal": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 9999999.99
                },
                "notes": {
                  "type": "string",
                  "maxLength": 500
                }
              },
              "required": [
                "product",
                "quantity",
                "price",
                "subtotal"
              ],
              "additionalProperties": false
            },
            "minItems": 1
          },
          "totalAmount": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999.99
          },
          "discount": {
            "type": "number",
            "minimum": 0,
            "maximum": 9999999.99
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "cash",
              "card",
              "credit_card",
              "debit_card",
              "transfer",
              "mobile_payment",
              "other"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "paid",
              "pending",
              "partial",
              "cancelled"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "pending",
              "cancelled"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 1000
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          },
          "cashier": {
            "$ref": "#/definitions/SaleCreateRequest/properties/customer"
          },
          "createdBy": {
            "$ref": "#/definitions/SaleCreateRequest/properties/customer"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          }
        },
        "required": [
          "items",
          "totalAmount",
          "paymentMethod"
        ],
        "additionalProperties": false
      },
...
      "CustomerCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "code": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "pattern": "^[\\d\\s\\-+()]+$",
            "minLength": 8,
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "maxLength": 254
          },
          "address": {
            "type": "string"
          },
          "idCardNumber": {
            "type": "string",
            "pattern": "^[A-Z][12]\\d{8}$"
          },
          "birthdate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "membershipLevel": {
            "type": "string",
            "enum": [
              "regular",
              "silver",
              "gold",
              "platinum"
            ]
          },
          "medicalHistory": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
...
    }
  }
}